Crate bevy_scene

Modules

prelude

serde

Structs

DynamicEntity

A reflection-powered serializable representation of an entity and its components.

DynamicScene

A collection of serializable dynamic entities, each with its own run-time defined set of components. To spawn a dynamic scene, you can use either:

DynamicSceneBundle

A component bundle for a DynamicScene root.

InstanceId

Scene

To spawn a scene, you can use either:

SceneBundle

A component bundle for a Scene root.

SceneInstance

InstanceId of a spawned scene. It can be used with the SceneSpawner to interact with the spawned scene.

SceneLoader

ScenePlugin

SceneSpawner

Enums

SceneSpawnError

Functions

scene_spawner

System that will spawn scenes from SceneBundle.

scene_spawner_system

serialize_ron

Serialize a given Rust data structure into rust object notation (ron).