bevy_scene
Enum SceneFilter

pub enum SceneFilter {
    Unset,
    Allowlist(HashSet<TypeId>),
    Denylist(HashSet<TypeId>),
}

A filter used to control which types can be added to a DynamicScene.

This scene filter can be used more generically to represent a filter for any given type; however, note that its intended usage with DynamicScene only considers components and resources. Adding types that are not a component or resource will have no effect when used with DynamicScene.

Variants

Unset

Represents an unset filter.

This is the equivalent of an empty Denylist or an Allowlist containing every typeโ essentially, all types are permissible.

Allowing a type will convert this filter to an Allowlist. Similarly, denying a type will convert this filter to a Denylist.

Allowlist(HashSet<TypeId>)

Contains the set of permitted types by their TypeId.

Types not contained within this set should not be allowed to be saved to an associated DynamicScene.

Denylist(HashSet<TypeId>)

Contains the set of prohibited types by their TypeId.

Types contained within this set should not be allowed to be saved to an associated DynamicScene.

Implementations

impl SceneFilter

pub fn allow_all() -> Self

Creates a filter where all types are allowed.

This is the equivalent of creating an empty Denylist.

pub fn deny_all() -> Self

Creates a filter where all types are denied.

This is the equivalent of creating an empty Allowlist.

pub fn allow<T: Any>(self) -> Self

Allow the given type, T.

If this filter is already set as a Denylist, then the given type will be removed from the denied set.

If this filter is Unset, then it will be completely replaced by a new Allowlist.

pub fn allow_by_id(self, type_id: TypeId) -> Self

Allow the given type.

If this filter is already set as a Denylist, then the given type will be removed from the denied set.

If this filter is Unset, then it will be completely replaced by a new Allowlist.

pub fn deny<T: Any>(self) -> Self

Deny the given type, T.

If this filter is already set as an Allowlist, then the given type will be removed from the allowed set.

If this filter is Unset, then it will be completely replaced by a new Denylist.

pub fn deny_by_id(self, type_id: TypeId) -> Self

Deny the given type.

If this filter is already set as an Allowlist, then the given type will be removed from the allowed set.

If this filter is Unset, then it will be completely replaced by a new Denylist.

pub fn is_allowed<T: Any>(&self) -> bool

Returns true if the given type, T, is allowed by the filter.

If the filter is Unset, this will always return true.

pub fn is_allowed_by_id(&self, type_id: TypeId) -> bool

Returns true if the given type is allowed by the filter.

If the filter is Unset, this will always return true.

pub fn is_denied<T: Any>(&self) -> bool

Returns true if the given type, T, is denied by the filter.

If the filter is Unset, this will always return false.

pub fn is_denied_by_id(&self, type_id: TypeId) -> bool

Returns true if the given type is denied by the filter.

If the filter is Unset, this will always return false.

pub fn iter(&self) -> Box<dyn ExactSizeIterator<Item = &TypeId> + '_>

Returns an iterator over the items in the filter.

If the filter is Unset, this will return an empty iterator.

pub fn len(&self) -> usize

Returns the number of items in the filter.

If the filter is Unset, this will always return a length of zero.

pub fn is_empty(&self) -> bool

Returns true if there are zero items in the filter.

If the filter is Unset, this will always return true.