bevy_scene
Struct DynamicScene

pub struct DynamicScene {
    pub resources: Vec<Box<dyn PartialReflect>>,
    pub entities: Vec<DynamicEntity>,
}

A collection of serializable resources and dynamic entities.

Each dynamic entity in the collection contains its own run-time defined set of components. To spawn a dynamic scene, you can use either:

SceneSpawner::spawn_dynamic

adding the DynamicSceneRoot component to an entity.

using the DynamicSceneBuilder to construct a DynamicScene from World.

Fields
resources: Vec<Box<dyn PartialReflect>>

Resources stored in the dynamic scene.

entities: Vec<DynamicEntity>

Entities contained in the dynamic scene.

Implementations

impl DynamicScene

pub fn from_scene(scene: &Scene) -> Self

Create a new dynamic scene from a given scene.

pub fn from_world(world: &World) -> Self

Create a new dynamic scene from a given world.

pub fn write_to_world_with( &self, world: &mut World, entity_map: &mut EntityHashMap<Entity>, type_registry: &AppTypeRegistry, ) -> Result<(), SceneSpawnError>

Write the resources, the dynamic entities, and their corresponding components to the given world.

This method will return a SceneSpawnError if a type either is not registered in the provided AppTypeRegistry resource, or doesn’t reflect the Component or Resource trait.

pub fn write_to_world( &self, world: &mut World, entity_map: &mut EntityHashMap<Entity>, ) -> Result<(), SceneSpawnError>

Write the resources, the dynamic entities, and their corresponding components to the given world.

This method will return a SceneSpawnError if a type either is not registered in the world’s AppTypeRegistry resource, or doesn’t reflect the Component trait.

pub fn serialize(&self, registry: &TypeRegistry) -> Result<String, Error>

Serialize this dynamic scene into the official Bevy scene format (.scn / .scn.ron).

The Bevy scene format is based on Rusty Object Notation (RON). It describes the scene in a human-friendly format. To deserialize the scene, use the SceneLoader.