Struct bevy_scene::Scene

pub struct Scene {
    pub world: World,
}

To spawn a scene, you can use either:

SceneSpawner::spawn

adding the SceneBundle to an entity

adding the Handle<Scene> to an entity (the scene will only be visible if the entity already has Transform and GlobalTransform components)

Fields
world: World
Implementations

impl Scene

pub fn new(world: World) -> Self

pub fn from_dynamic_scene( dynamic_scene: &DynamicScene, type_registry: &AppTypeRegistry ) -> Result<Scene, SceneSpawnError>

Create a new scene from a given dynamic scene.

pub fn clone_with( &self, type_registry: &AppTypeRegistry ) -> Result<Scene, SceneSpawnError>

Clone the scene.

This method will return a SceneSpawnError if a type either is not registered in the provided AppTypeRegistry or doesn’t reflect the Component trait.

pub fn write_to_world_with( &self, world: &mut World, type_registry: &AppTypeRegistry ) -> Result<InstanceInfo, SceneSpawnError>

Write the entities and their corresponding components to the given world.

This method will return a SceneSpawnError if a type either is not registered in the provided AppTypeRegistry or doesn’t reflect the Component trait.