bevy_scene
Struct SceneSpawner

pub struct SceneSpawner { /* private fields */ }

Handles spawning and despawning scenes in the world, either synchronously or batched through the scene_spawner_system.

Synchronous methods: (Scene operations will take effect immediately)

spawn_dynamic_sync

spawn_sync

despawn_sync

despawn_instance_sync

update_spawned_scenes

spawn_queued_scenes

despawn_queued_scenes

despawn_queued_instances

Deferred methods: (Scene operations will be processed when the scene_spawner_system is run)

spawn_dynamic

spawn_dynamic_as_child

spawn

spawn_as_child

despawn

despawn_instance

Implementations

impl SceneSpawner

pub fn spawn_dynamic( &mut self, id: impl Into<Handle<DynamicScene>>, ) -> InstanceId

Schedule the spawn of a new instance of the provided dynamic scene.

pub fn spawn_dynamic_as_child( &mut self, id: impl Into<Handle<DynamicScene>>, parent: Entity, ) -> InstanceId

Schedule the spawn of a new instance of the provided dynamic scene as a child of parent.

pub fn spawn(&mut self, id: impl Into<Handle<Scene>>) -> InstanceId

Schedule the spawn of a new instance of the provided scene.

pub fn spawn_as_child( &mut self, id: impl Into<Handle<Scene>>, parent: Entity, ) -> InstanceId

Schedule the spawn of a new instance of the provided scene as a child of parent.

pub fn despawn(&mut self, id: impl Into<AssetId<DynamicScene>>)

Schedule the despawn of all instances of the provided dynamic scene.

pub fn despawn_instance(&mut self, instance_id: InstanceId)

Schedule the despawn of a scene instance, removing all its entities from the world.

pub fn despawn_sync( &mut self, world: &mut World, id: impl Into<AssetId<DynamicScene>>, ) -> Result<(), SceneSpawnError>

Immediately despawns all instances of a dynamic scene.

pub fn despawn_instance_sync( &mut self, world: &mut World, instance_id: &InstanceId, )

Immediately despawns a scene instance, removing all its entities from the world.

pub fn spawn_dynamic_sync( &mut self, world: &mut World, id: impl Into<AssetId<DynamicScene>>, ) -> Result<InstanceId, SceneSpawnError>

Immediately spawns a new instance of the provided dynamic scene.

pub fn spawn_sync( &mut self, world: &mut World, id: impl Into<AssetId<Scene>>, ) -> Result<InstanceId, SceneSpawnError>

Immediately spawns a new instance of the provided scene.

pub fn update_spawned_scenes( &mut self, world: &mut World, scene_ids: &[AssetId<DynamicScene>], ) -> Result<(), SceneSpawnError>

Iterate through all instances of the provided scenes and update those immediately.

Useful for updating already spawned scene instances after their corresponding scene has been modified.

pub fn despawn_queued_scenes( &mut self, world: &mut World, ) -> Result<(), SceneSpawnError>

Immediately despawns all scenes scheduled for despawn by despawning their instances.

pub fn despawn_queued_instances(&mut self, world: &mut World)

Immediately despawns all scene instances scheduled for despawn.

pub fn spawn_queued_scenes( &mut self, world: &mut World, ) -> Result<(), SceneSpawnError>

Immediately spawns all scenes scheduled for spawn.

pub fn instance_is_ready(&self, instance_id: InstanceId) -> bool

Check that an scene instance spawned previously is ready to use

pub fn iter_instance_entities( &self, instance_id: InstanceId, ) -> impl Iterator<Item = Entity> + '_

Get an iterator over the entities in an instance, once it’s spawned.

Before the scene is spawned, the iterator will be empty. Use Self::instance_is_ready to check if the instance is ready.