Crate bevy_scene

Provides scene definition, instantiation and serialization/deserialization.

Scenes are collections of entities and their associated components that can be instantiated or removed from a world to allow composition. Scenes can be serialized/deserialized, for example to save part of the world state to a file.

Re-exports

pub use bevy_asset::ron;

Modules

prelude

The scene prelude.

serde

serde serialization and deserialization implementation for Bevy scenes.

Structs

DynamicEntity

A reflection-powered serializable representation of an entity and its components.

DynamicScene

A collection of serializable resources and dynamic entities.

DynamicSceneBuilder

A DynamicScene builder, used to build a scene from a World by extracting some entities and resources.

DynamicSceneBundleDeprecated

A component bundle for a DynamicScene root.

DynamicSceneRoot

Adding this component will spawn the scene as a child of that entity. Once it’s spawned, the entity will have a SceneInstance component.

InstanceId

Unique id identifying a scene instance.

InstanceInfo

Information about a scene instance.

Scene

A composition of World objects.

SceneBundleDeprecated

A component bundle for a Scene root.

SceneInstance

InstanceId of a spawned scene. It can be used with the SceneSpawner to interact with the spawned scene.

SceneInstanceReady

Triggered on a scene’s parent entity when crate::SceneInstance becomes ready to use.

SceneLoader

Asset loader for a Bevy dynamic scene (.scn / .scn.ron).

ScenePlugin

Plugin that provides scene functionality to an App.

SceneRoot

Adding this component will spawn the scene as a child of that entity. Once it’s spawned, the entity will have a SceneInstance component.

SceneSpawner

Handles spawning and despawning scenes in the world, either synchronously or batched through the scene_spawner_system.

Enums

SceneFilter

A filter used to control which types can be added to a DynamicScene.

SceneLoaderError

Possible errors that can be produced by SceneLoader

SceneSpawnError

Errors that can occur when spawning a scene.

Functions

scene_spawner

System that will spawn scenes from the SceneRoot and DynamicSceneRoot components.

scene_spawner_system

System that handles scheduled scene instance spawning and despawning through a SceneSpawner.

serialize_ron

Serialize a given Rust data structure into rust object notation (ron).