Struct bevy_scene::DynamicScene

pub struct DynamicScene {
    pub resources: Vec<Box<dyn Reflect>>,
    pub entities: Vec<DynamicEntity>,
}

A collection of serializable resources and dynamic entities.

Each dynamic entity in the collection contains its own run-time defined set of components. To spawn a dynamic scene, you can use either:

SceneSpawner::spawn_dynamic

adding the DynamicSceneBundle to an entity

adding the Handle<DynamicScene> to an entity (the scene will only be visible if the entity already has Transform and GlobalTransform components)

Fields
resources: Vec<Box<dyn Reflect>>entities: Vec<DynamicEntity>
Implementations

impl DynamicScene

pub fn from_scene(scene: &Scene) -> Self

Create a new dynamic scene from a given scene.

pub fn from_world(world: &World) -> Self

Create a new dynamic scene from a given world.

pub fn write_to_world_with( &self, world: &mut World, entity_map: &mut EntityMap, type_registry: &AppTypeRegistry ) -> Result<(), SceneSpawnError>

Write the resources, the dynamic entities, and their corresponding components to the given world.

This method will return a SceneSpawnError if a type either is not registered in the provided AppTypeRegistry resource, or doesn’t reflect the Component or Resource trait.

pub fn write_to_world( &self, world: &mut World, entity_map: &mut EntityMap ) -> Result<(), SceneSpawnError>

Write the resources, the dynamic entities, and their corresponding components to the given world.

This method will return a SceneSpawnError if a type either is not registered in the world’s AppTypeRegistry resource, or doesn’t reflect the Component trait.

pub fn serialize_ron(&self, registry: &TypeRegistryArc) -> Result<String, Error>

Serialize this dynamic scene into rust object notation (ron).