Struct bevy_scene::SceneSpawner

pub struct SceneSpawner { /* private fields */ }

Implementations

impl SceneSpawner

pub fn spawn_dynamic( &mut self, scene_handle: Handle<DynamicScene> ) -> InstanceId

pub fn spawn_dynamic_as_child( &mut self, scene_handle: Handle<DynamicScene>, parent: Entity ) -> InstanceId

pub fn spawn(&mut self, scene_handle: Handle<Scene>) -> InstanceId

pub fn spawn_as_child( &mut self, scene_handle: Handle<Scene>, parent: Entity ) -> InstanceId

pub fn despawn(&mut self, scene_handle: Handle<DynamicScene>)

pub fn despawn_instance(&mut self, instance_id: InstanceId)

pub fn despawn_sync( &mut self, world: &mut World, scene_handle: Handle<DynamicScene> ) -> Result<(), SceneSpawnError>

pub fn despawn_instance_sync( &mut self, world: &mut World, instance_id: &InstanceId )

pub fn spawn_dynamic_sync( &mut self, world: &mut World, scene_handle: &Handle<DynamicScene> ) -> Result<(), SceneSpawnError>

pub fn spawn_sync( &mut self, world: &mut World, scene_handle: Handle<Scene> ) -> Result<InstanceId, SceneSpawnError>

pub fn update_spawned_scenes( &mut self, world: &mut World, scene_handles: &[Handle<DynamicScene>] ) -> Result<(), SceneSpawnError>

pub fn despawn_queued_scenes( &mut self, world: &mut World ) -> Result<(), SceneSpawnError>

pub fn despawn_queued_instances(&mut self, world: &mut World)

pub fn spawn_queued_scenes( &mut self, world: &mut World ) -> Result<(), SceneSpawnError>

pub fn instance_is_ready(&self, instance_id: InstanceId) -> bool

Check that an scene instance spawned previously is ready to use

pub fn iter_instance_entities( &self, instance_id: InstanceId ) -> impl Iterator<Item = Entity> + '_

Get an iterator over the entities in an instance, once it’s spawned.

Before the scene is spawned, the iterator will be empty. Use Self::instance_is_ready to check if the instance is ready.